fix(ApiController, CoreController, PanelController): implement transformer logic for data handling
- Added a fallback transformer method in ApiController to handle collections when no specific resource class is defined.
- Introduced new transformer methods in CoreController to facilitate data transformation for paginated collections and individual resources.
- Refactored getJSONData method in PanelController to streamline data retrieval and transformation process.
- Removed redundant transformer methods from PanelController to maintain cleaner code structure.
